add edit + delete mycontact 22.11.21

+++ src/main/java/com/example/mycontact/entity/Contact.java
@@ -11,7 +11,7 @@ public class Contact {
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
     @Column(name="id",nullable = false)
-    private Integer id;
+    private Long id;
 
     @NotEmpty
     @Column(name="name",nullable = false)
@@ -27,18 +27,18 @@ public class Contact {
     public Contact() {
     }
 
-    public Contact(Integer id, String name, String email, String phone) {
+    public Contact(long id, String name, String email, String phone) {
         this.id = id;
         this.name = name;
         this.email = email;
         this.phone = phone;
     }
 
-    public Integer getId() {
+    public Long getId() {
         return id;
     }
 
-    public void setId(Integer id) {
+    public void setId(Long id) {
         this.id = id;
     }
