+++ src/main/java/com/example/mycontact/repository/ContactRepository.java
@@ -7,6 +7,6 @@ import org.springframework.stereotype.Repository;
 import java.util.List;
 
 @Repository
-public interface ContactRepository extends CrudRepository<Contact,Integer> {
+public interface ContactRepository extends CrudRepository<Contact,Long> {
     List<Contact> findByNameContaining(String term);
 }
